Add Tailwind config
Typically, this file is called `tailwind.js`, however, following Sage's
naming convention for config files, `tailwind.config.js` is more
suitable. Tailwind also recommends creating this file in the project's
root, others add it to `scripts/`, but I prefer to place it with the
`styles/` since — though it is JS — it is the source for the majority of
the styles. A strong case could be made for `build/` to place it
alongside `postcss.config.js`, since Tailwind is a PostCSS plugin.