Add initial take on Tailwind base styles
`main.scss` will most likely stay the same. I added the comments from
the their recommended stylesheet structure example since they are
informative: https://tailwindcss.com/docs/installation#3-use-tailwind-in-your-css

As well this adds a Tailwind interpretation of Sage's base styles. Of
course, since the markup is remaining in tact, I am using Tailwind's
`@apply` to take advantage of the existing selectors in the markup. A
stylistic choice I made here is to omit the period at the beginning of
the class names being applied (i.e. `@apply bg-primary text-white`
instead of `@apply .bg-primary .text-white` similar to `@extend` in
Sass). This is because when extracting components in Tailwind, it is
more ergonomic to grab all the selectors in a class attribute and chuck
them behind an `@apply` vice versa then to prepend a period everytime.