Add Tailwind webpack.config.preset.js
This "fixes" a known issue when using Tailwind with Sass
(bholloway/resolve-url-loader#28). It removes the `resolve-url-loader`
from the loaders for SCSS files and disables source maps. Without this
change there would be a persistent warning when running `yarn start`.

This change requires a change to the webpack.config.js in roots/sage.
The final export needs to be:

```
module.exports = merge.smartStrategy({
  'module.loaders': 'replace'
})(webpackConfig, desire(`${__dirname}/webpack.config.preset`));
```